Connected filters to accordion charts
