+++ src/pages/bioclimatic/component.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 import { LayerManager, Layer } from 'layer-manager/dist/components';
 import { PluginMapboxGl } from 'layer-manager';
 import sortBy from 'lodash/sortBy';
@@ -11,67 +12,50 @@ import LayerToggle from 'components/map/controls/layer-toggle';
 import { useBiovars } from 'graphql/queries';
 
 import layers from 'layers.json';
-import styles from './styles.scss';
 
-function BioClimaticPage() {
+function BioClimaticPage(props) {
+  const { config, filters } = props;
+  const { startYear, endYear, scenario, parsedScenarios } = filters;
+  const parsedScenario = parsedScenarios && parsedScenarios.find(s => s.value === scenario).label;
+
   const [activeLayers, setActiveLayers] = useState(layers.map(l => ({ ...l, active: true })));
+  const { fetching, data: biovars, error } = useBiovars();
 
-  const { fetching, data, error } = useBiovars();
-
+  // TODO: change to graphQL query based on scenario, biovar (item), and country
   const mockData = [
-    { name: '2020', 'Business as usual': 3 },
-    { name: '2030', 'Business as usual': 3.5 },
-    { name: '2040', 'Business as usual': 6 },
-    { name: '2050', 'Business as usual': 6.2 },
-    { name: '2060', 'Business as usual': 11 },
-    { name: '2070', 'Business as usual': 9 },
-    { name: '2080', 'Business as usual': 13 },
-    { name: '2090', 'Business as usual': 18 }
+    { name: '2020', [scenario]: 3 },
+    { name: '2030', [scenario]: 3.5 },
+    { name: '2040', [scenario]: 6 },
+    { name: '2050', [scenario]: 6.2 },
+    { name: '2060', [scenario]: 11 },
+    { name: '2070', [scenario]: 9 },
+    { name: '2080', [scenario]: 13 },
+    { name: '2090', [scenario]: 18 }
   ];
 
-  const config = {
-    lines: [
-      {
-        key: 'Business as usual',
-        color: styles.colorPink
-      }
-    ],
-    areas: [
-      {
-        key: 'Business as usual',
-        color: styles.colorPink
-      }
-    ],
-    yAxis: {
-      domain: [0, 20],
-      unit: '°C',
-      ticks: [0, 5, 10, 15, 20],
-      customTick: true,
-      axisLine: false
-    },
-    xAxis: {
-      padding: { left: -30, right: -30 }
-    },
-    grid: {
-      vertical: false
-    },
-    height: 300
-  };
+  const filterData = data =>
+    data.filter(d => Number(d.name) >= Number(startYear) && Number(d.name) <= Number(endYear));
 
   return (
     <div className="c-bioclimatic l-page">
-      <Filters />
+      {/* eslint-disable-next-line react/jsx-props-no-spreading */}
+      <Filters {...filters} />
       <div className="content">
         {fetching && <p>Loading...</p>}
         {error && <p>Error retrieving the data</p>}
         {!fetching && !error && (
           <div className="bioclimatic-chart">
             <Accordion
-              items={sortBy(data.biovars, 'key').map((bv, i) => ({
+              items={sortBy(biovars.biovars, 'key').map((bv, i) => ({
                 title: `BIO ${i + 1} = ${bv.name}`,
                 key: bv.key,
-                data: mockData,
-                config
+                data: filterData(mockData),
+                config,
+                metadata: {
+                  dataset: bv.name.replace(/ *\([^)]*\) */g, ''),
+                  model: parsedScenario,
+                  unit: '°C' // TODO: change
+                }
               }))}
             />
           </div>
@@ -106,5 +90,10 @@ function BioClimaticPage() {
     </div>
   );
 }
+
+BioClimaticPage.propTypes = {
+  config: PropTypes.object,
+  filters: PropTypes.object
+};
 
 export default BioClimaticPage;
